policy files edit

+++ app/Http/Controllers/Api/BrandController.php
@@ -8,6 +8,8 @@ use App\Http\Requests\UpdateBrandRequest;
 use App\Interfaces\BrandRepositoryInterface;
 use Illuminate\Http\Request;
 use App\Models\Brand;
+use Illuminate\Support\Facades\Gate;
+
 
 class BrandController extends Controller
 {
@@ -20,6 +22,7 @@ class BrandController extends Controller
 
     public function index(Request $request)
     {
+        Gate::authorize("viewAny", Brand::class);
          // Request'ten page ve perPage parametrelerini alıyoruz. 
          $page = $request->input('page', 1); 
          $perPage = $request->input('per_page', 15); 
@@ -36,27 +39,28 @@ class BrandController extends Controller
     public function show($id)
     {
         $brand = $this->brandRepository->find($id);
-        if ($brand) {
-            return response()->json([
-                'success' => true,
-                'data' => $brand,
-                'errors' => null,
-                'message' => null,
-            ], 200);
-        } else {
+        if (!$brand) {
             return response()->json([
                 'success' => false,
                 'data' => null,
                 'errors' => 'Brand not found.',
                 'message' => 'The requested brand does not exist.',
             ], 400);
+        } else {
+            Gate::authorize('view', Brand::class);
+            return response()->json([
+                'success' => true,
+                'data' => $brand,
+                'errors' => null,
+                'message' => null,
+            ], 200);
         }
     }
 
     public function store(StoreBrandRequest $request)
     {
+        Gate::authorize('create', Brand::class);
         $validatedData = $request->validated();
-
         try {
             $brand = $this->brandRepository->create($validatedData);
             return response()->json([
@@ -78,6 +82,18 @@ class BrandController extends Controller
     public function update(UpdateBrandRequest $request, $id)
     {
         $validatedData = $request->validated();
+
+        $brand = $this->brandRepository->find($id);
+        if (!$brand) {
+            return response()->json([
+                'success' => false,
+                'data' => null,
+                'errors' => 'Brand not found.',
+                'message' => 'The requested brand does not exist.',
+            ], 400);
+        } 
+
+        Gate::authorize('update', $brand);
 
         $brand = $this->brandRepository->update($id, $validatedData);
         if ($brand) 
@@ -104,6 +120,18 @@ class BrandController extends Controller
 
     public function destroy($id)
     {
+        $brand = $this->brandRepository->find($id);
+        if (!$brand) {
+            return response()->json([
+                'success' => false,
+                'data' => null,
+                'errors' => 'Brand not found.',
+                'message' => 'The requested brand does not exist.',
+            ], 400);
+        } 
+
+        Gate::authorize('delete', $brand);
+
         $brand = $this->brandRepository->delete($id);
         if ($brand) {
            return response()->json([
