+++ app/Http/Controllers/Api/PaymentStatusController.php
@@ -4,8 +4,10 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StatusRequest;
+use App\Models\PaymentStatus;
 use App\Repositories\PaymentStatusRepository;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class PaymentStatusController extends Controller
 {
@@ -18,6 +20,7 @@ class PaymentStatusController extends Controller
 
     public function index()
     {
+        Gate::authorize('viewAny', PaymentStatus::class);
         $statuses = $this->statusRepository->getAllStatuses();
         return response()->json([
             'success' => true,
@@ -30,24 +33,28 @@ class PaymentStatusController extends Controller
     public function show($id)
     {
         $status = $this->statusRepository->getStatusById($id);
-        if ($status) {
+        if (!$status) {
             return response()->json([
-                'success' => true,
-                'data' => $status,
-                'errors' => null,
+                'success' => false,
+                'data' => null,
+                'errors' => 'Payment status not found.',
                 'message' => null,
-            ]);
+            ], 400);
         }
+
+        Gate::authorize('view', $status);
+
         return response()->json([
-            'success' => false,
-            'data' => null,
-            'errors' => 'Payment status not found.',
+            'success' => true,
+            'data' => $status,
+            'errors' => null,
             'message' => null,
-        ], 400);
+        ],200);
     }
 
     public function store(StatusRequest $request)
     {
+        Gate::authorize('create',PaymentStatus::class);
         $data = $request->validated();
         $status = $this->statusRepository->createStatus($data);
         return response()->json([
@@ -61,6 +68,18 @@ class PaymentStatusController extends Controller
     public function update(StatusRequest $request, $id)
     {
         $data = $request->validated();
+        $status = $this->statusRepository->getStatusById($id);
+        if (!$status) {
+            return response()->json([
+                'success' => false,
+                'data' => null,
+                'errors' => 'Payment status not found.',
+                'message' => null,
+            ], 400);
+        }
+
+        Gate::authorize('update',$status);
+        
         $status = $this->statusRepository->updateStatus($id, $data);
         if ($status) {
             return response()->json([
@@ -73,13 +92,14 @@ class PaymentStatusController extends Controller
         return response()->json([
             'success' => false,
             'data' => null,
-            'errors' => 'Payment status not found.',
+            'errors' => 'Payment status güncellenemedi !',
             'message' => null,
         ], 400);
     }
 
     public function destroy($id)
     {
+        Gate::authorize('delete');
         if ($this->statusRepository->deleteStatus($id)) {
             return response()->json([
                 'success' => true,
